pyproject.toml
[tool.poetry]
name = "terminologyproject"
version = "0.1.0"
description = ""
authors = ["German Legkovoy <147488090+legkovoyg@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
